requirements.txt
attrs==23.2.0
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
filelock==3.13.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==0.10.0
json5==0.9.24
lxml==5.1.0
numpy==1.26.4
packaging==24.0
parsel==1.9.0
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
python-dotenv==1.0.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
scrapeops_scrapy==0.5.4
Scrapy==2.11.1
scrapy-playwright==0.0.34
service-identity==24.1.0
setuptools==69.1.1
six==1.16.0
tld==0.13
tldextract==5.1.1
Twisted==24.3.0
typing_extensions==4.10.0
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.2
